requirements.txt
# --index-url https://download.pytorch.org/whl/cu118
torch

notebook
opencv-python
matplotlib
numpy
pandas
plotly
pyyaml
tqdm
python-dotenv
isort
ultralytics
py-markdown-table
onnx
onnx-simplifier
torchinfo

# tensorflow
# tflite-support
# onnx2tf
# sng4onnx
# nvidia-pyindex
# onnx-graphsurgeon

# ultralytics issue
# lap